Fixed architecture problems
