+++ FrontEnd/src/app/auth/login/Login.tsx
@@ -2,56 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import './Login.scss';
 import LoginPic from '../../../assets/images/LoginPic.jpg';
-
-const WelcomeMessege = () => {
-    return (
-        <div className="welcome-container">
-            <h1 className="first-title">Welcome to FixFlow</h1>
-            <h2 className="sub-title">
-                We are a new platform that makes your company a more organized workplace with our smart ticketing system
-            </h2>
-        </div>
-    );
-};
-
-const MotivationSentences = () => {
-    const messages = [
-        'Unlock a world of possibilities.',
-        "Discover what you've been missing.",
-        'Ready to elevate your experience?',
-        'Streamline your life, one click at a time.',
-        'Experience perfect order.',
-        'Bring clarity to your chaos.',
-        'Your organized future starts here.',
-        'Join us and transform your workflow.',
-        'Sign up today and see the difference.',
-        "Get started now, it's free and easy.",
-    ];
-
-    const [index, setIndex] = useState(0);
-    const [isVisible, setIsVisible] = useState(true);
-
-    useEffect(() => {
-        const interval = setInterval(() => {
-            setIsVisible(false);
-
-            setTimeout(() => {
-                setIndex((prevIndex) => (prevIndex + 1) % messages.length);
-                setIsVisible(true);
-            }, 600);
-        }, 4000);
-
-        return () => clearInterval(interval);
-    }, [messages.length]);
-
-    return (
-        <div className={`Motivation-Sentences ${isVisible ? 'visible' : 'hidden'}`}>
-            <h3 style={{ color: 'black', fontSize: '30px' }} className="dynamic-text">
-                {messages[index]}
-            </h3>
-        </div>
-    );
-};
+import WelcomeMessage from './WelcomeMessage';
+import MotivationSentences from './MotivationSentences';
 
 const Picture = () => {
     return (
@@ -78,7 +30,7 @@ const Login = () => {
 
     return (
         <div className="login-container">
-            <WelcomeMessege />
+            <WelcomeMessage />
             <MotivationSentences />
             <main className="main-content-split">
                 <Picture />
