+++ FrontEnd/src/app/auth/login/WelcomeMessage.scss
@@ -1,0 +1,27 @@
+@use '../../../shared/styles/colors';
+
+.welcome-container {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+    gap: 25px;
+    margin-top: 20px;
+    margin-bottom: 40px;
+
+    .first-title {
+        color: colors.$neutral9;
+        font-size: 48px;
+        margin: 0;
+    }
+
+    .sub-title {
+        color: colors.$purple6;
+        width: 100%;
+        font-size: 30px;
+        text-align: center;
+        max-width: 1000px;
+        margin: 0 auto;
+    }
+}
