+++ FrontEnd/src/app/auth/login/WelcomeMessage.tsx
@@ -1,0 +1,11 @@
+const WelcomeMessage = () => {
+    return (
+        <div className="welcome-container">
+            <h1 className="first-title">Welcome to FixFlow</h1>
+            <h2 className="sub-title">
+                We are a new platform that makes your company a more organized workplace with our smart ticketing system
+            </h2>
+        </div>
+    );
+};
+export default WelcomeMessage;
